+++ playground/src/Pages/DialogPage.tsx
@@ -29,6 +29,10 @@ export const DialogPage: React.FC = () => {
     }, delay)
   }
 
+  const [dialogDisableClickAway, setDialogDisableClickAway] = React.useState<boolean>(false)
+  const showDialogDisableClickAway = (): void => setDialogDisableClickAway(true)
+  const hideDialogDisableClickAway = (): void => setDialogDisableClickAway(false)
+
   return (
     <>
       <Header><Code>{'<Dialog>'}</Code></Header>
@@ -63,11 +67,6 @@ setOpen: React.Dispatch<React.SetStateAction<boolean>> | (() => void)`}
         <Code>{'<Dialog>'}</Code> also accepts the following optional props:
       </Paragraph>
       <CodeBlock>{`/**
- * Optional header content for the Dialog
- */
-header?: React.ReactNode
-
-/**
  * An array of actions to generate \`<Button>\`s at the bottom of the Dialog
  */
 actions?: {
@@ -79,8 +78,31 @@ actions?: {
   primary?: boolean
   /** Optional overrides for the \`<Button>\`'s props */
   buttonProps?: ButtonProps
-}[]`}
+}[]
+
+/**
+ * If true, the user cannot click away from the dialog
+ */
+disableClickAway?: boolean
+
+/**
+ * Optional header content for the Dialog
+ */
+header?: React.ReactNode
+
+`}
       </CodeBlock>
+      <Header as='h3' id='disableclickaway'>Notes on usage of <Code>disableClickAway?: boolean</Code></Header>
+      <Paragraph>
+        The <Code>disableClickAway</Code> prop creates a <Code>{'<Dialog>'}</Code> that cannot be closed by the user
+        by clicking away from the dialog or by hitting the <Code>esc</Code> key. Because it traps the user, this prop
+        has no effect if <Code>actions</Code> is undefined&mdash;that is, you cannot disable clickaway behavior on
+        a <Code>{'<Dialog>'}</Code> with no buttons.
+      </Paragraph>
+      <Paragraph>
+        It is absolutely essential to properly handle closing the dialog in the <Code>onClick</Code> functions of your
+        action buttons when setting <Code>disableClickAway</Code> to <Code>true</Code>.
+      </Paragraph>
       <Header as='h2' id='customization'>Customization</Header>
       <Paragraph>
         See <AccessibleHashlink to={'/start#customization'}>Customization</AccessibleHashlink> discussion on the
@@ -90,6 +112,7 @@ actions?: {
       <Paragraph as ='div'>
         <Button onClick={showBasicDialog}>Show basic dialog</Button>
         <Dialog
+          // disableClickAway
           open={basicDialog}
           setOpen={setBasicDialog}
         >
@@ -100,8 +123,7 @@ actions?: {
             you can dismiss with <Code>esc</Code> or by clicking the backdrop.
           </Paragraph>
         </Dialog>
-      </Paragraph>
-      <Paragraph as='div'>
+      </Paragraph><Paragraph as='div'>
         <Button onClick={showDialogWithActions}>Show dialog with actions</Button>
         <Dialog
           actions={[
@@ -134,6 +156,29 @@ actions?: {
         >
           Deleted
         </Notification>
+      </Paragraph>
+      <Paragraph as='div'>
+        <Button onClick={showDialogDisableClickAway}>Show dialog with disableClickAway</Button>
+        <Dialog
+          disableClickAway
+          actions={[
+            {
+              label: 'Close',
+              primary: true,
+              onClick: hideDialogDisableClickAway,
+            },
+          ]}
+          open={dialogDisableClickAway}
+          setOpen={setDialogDisableClickAway}
+          header={'You can\'t click away.'}
+        >
+          <Paragraph>
+            This <Code>{'<Dialog>'}</Code> has the prop <Code>disableClickAway</Code> set to true.
+          </Paragraph>
+          <Paragraph>
+            To dismiss this dialog, the user must click the "close" button below.
+          </Paragraph>
+        </Dialog>
       </Paragraph>
     </>
   )}