+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `color?: Colors` prop added to `<Notification>`
+- React hook `useClickAway`
+- `<Notification>` clickaway behavior
+- `<Dialog>` prop `disableClickAway`
 ### Changed
 - Give `<Button>` same size/padding/border/margin as inputs so they line up nicely side by side
 - Change how `complimentaryColor` works